Fix issue [#4] - HashMap's empty LinkedLists
Fix issue on __delitem__ method where it was not substituting
empty LinkedLists for None items.

Items with the same hashed key value were forcefully inserted into the
emtpy LinkedList, even though it was the only item inside the bucket, thus
creating an unnecessary number of LinkedLists on the hashmap over time.